feat!: rewrite to JS
BREAKING CHANGE: now uses ES modules
