+++ .eslintrc.json
@@ -1,4 +1,8 @@
 {
+  "env": {
+    "es6": true,
+    "node": true
+  },
   "extends": [
     "eslint:recommended",
     "plugin:@typescript-eslint/recommended",
@@ -6,11 +10,5 @@
   ],
   "parserOptions": {
     "project": "./tsconfig.json"
-  },
-  "rules": {
-    "@typescript-eslint/strict-boolean-expressions": "error",
-    "eqeqeq": ["error", "smart"],
-    "no-else-return": "error",
-    "no-implicit-coercion": "error"
   }
 }
