+++ .gitignore
@@ -1,3 +1,3 @@
 /node_modules/
-/lib/
-/coverage/
+/lib/*.d.ts
+lcov.info
