+++ tests/media/pink.png
