+++ tsconfig.json
@@ -1,12 +1,12 @@
 {
   "compilerOptions": {
-    "declaration": true,
-    "esModuleInterop": true,
-    "lib": ["ES2020"],
-    "module": "CommonJS",
-    "moduleResolution": "Node",
-    "outDir": "lib",
+    "allowJs": true,
+    "checkJs": true,
+    "lib": ["ES2023"],
+    "module": "Node16",
+    "moduleResolution": "Node16",
+    "noEmit": true,
     "strict": true,
-    "target": "ES2018"
+    "target": "ES2023"
   }
 }
