--- types/is-mime.d.ts
@@ -1,4 +1,0 @@
-declare module "is-mime" {
-  import type { PassThrough } from "stream"
-  export function checkStream(mimes: string[]): PassThrough
-}
